+++ src/init.c
@@ -51,7 +51,7 @@ extern r_obj* ffi_as_location(r_obj*, r_obj*, r_obj*, r_obj*, r_obj*, r_obj*, r_
 extern r_obj* ffi_slice(r_obj*, r_obj*, r_obj*);
 extern r_obj* ffi_init(r_obj*, r_obj*, r_obj*);
 extern r_obj* ffi_vec_chop(r_obj*, r_obj*, r_obj*);
-extern r_obj* ffi_list_unchop(r_obj*, r_obj*, r_obj*, r_obj*, r_obj*, r_obj*);
+extern r_obj* ffi_list_unchop(r_obj*, r_obj*, r_obj*, r_obj*, r_obj*, r_obj*, r_obj*, r_obj*);
 extern r_obj* ffi_vec_chop_seq(r_obj*, r_obj*, r_obj*, r_obj*);
 extern r_obj* ffi_slice_seq(r_obj*, r_obj*, r_obj*, r_obj*);
 extern r_obj* ffi_slice_rep(r_obj*, r_obj*, r_obj*);
@@ -235,7 +235,7 @@ static const R_CallMethodDef CallEntries[] = {
   {"ffi_slice",                                 (DL_FUNC) &ffi_slice, 3},
   {"ffi_init",                                  (DL_FUNC) &ffi_init, 3},
   {"ffi_vec_chop",                              (DL_FUNC) &ffi_vec_chop, 3},
-  {"ffi_list_unchop",                           (DL_FUNC) &ffi_list_unchop, 6},
+  {"ffi_list_unchop",                           (DL_FUNC) &ffi_list_unchop, 8},
   {"ffi_vec_chop_seq",                          (DL_FUNC) &ffi_vec_chop_seq, 4},
   {"ffi_slice_seq",                             (DL_FUNC) &ffi_slice_seq, 4},
   {"ffi_slice_rep",                             (DL_FUNC) &ffi_slice_rep, 3},