+++ NEWS.md
@@ -1,5 +1,7 @@
 # vctrs (development version)
 
+* `list_unchop()` has gained new `size` and `default` arguments (#TODO).
+
 * `vec_detect_complete(NULL)` now returns `logical()`, consistent with
   `vec_detect_missing(NULL)` (#1916).